Subject Creating a simple Rock Paper Scissors game
Problem
Creating a game from scratch in javascript.

Solution
Building it using what I have learned so far, using functions and
variables. Will have to also use the dev tools to be able to play it
in the console for now.

Note
Starting it by first breaking down the problem using pen and paper.
After breaking it down, start coding it, and go back and fourth
between coding and writing it out, as more problems need to be written
out and broken down into simpler problems.