Subject Created functions for the game.
Problem
Needed to create functions to be able to make the game.

Solution
created multiple functions, one to get a random number. To then use
the number in another function for it to choose either rock, paper,
or scissors.

Note
Using console.log to check at multiple steps to see what the code is
doing, so I can make sure it is working properly.