Subject Created the game to play five rounds.
Problem
Needed to create a game that plays five rounds of Rock Paper Scissors.
Needed to make a function that played five rounds.

Solution
created the game() function that called upon the other functions that
were created to select the computer choice, player choice and the one to
determine who won the game.